requirements.txt
pandas
requests
pyarrow
apache-airflow==2.7.3
pytest
